Rename py_wa_replication to wa_replication.
